Make TypeReferences a nested class of SourceFileWithTypeReferences
