+++ rewrite-core/src/main/java/org/openrewrite/SourceFileWithTypeReferences.java
@@ -15,9 +15,42 @@
  */
 package org.openrewrite;
 
-import org.openrewrite.internal.TypeReferences;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import org.openrewrite.trait.TypeReference;
+
+import java.util.*;
 
 public interface SourceFileWithTypeReferences extends SourceFile {
 
     TypeReferences getTypeReferences();
+
+    @RequiredArgsConstructor(access = AccessLevel.PRIVATE)
+    @Getter
+    class TypeReferences {
+        private final SourceFile sourceFile;
+        private final Set<TypeReference> typeReferences;
+
+        public Collection<TypeReference> findMatches(TypeReference.Matcher matcher) {
+            List<TypeReference> list = new ArrayList<>();
+            for (TypeReference ref : typeReferences) {
+                if (ref.matches(matcher)) {
+                    list.add(ref);
+                }
+            }
+            return list;
+        }
+
+        public static TypeReferences build(SourceFile sourceFile) {
+            Set<TypeReference> typeReferences = new HashSet<>();
+            ServiceLoader<TypeReference.Provider> loader = ServiceLoader.load(TypeReference.Provider.class);
+            loader.forEach(provider -> {
+                if (provider.isAcceptable(sourceFile)) {
+                    typeReferences.addAll(provider.getTypeReferences(sourceFile));
+                }
+            });
+            return new TypeReferences(sourceFile, typeReferences);
+        }
+    }
 }
